test: overlap AGY event observation with scenario flow
